--- src/main/java/Command.java
@@ -1,3 +1,0 @@
-public interface Command {
-    void execute(ItemList itemList, Ui ui, Storage storage);
-}
