+++ src/main/java/atlas/commands/ByeCommand.java
@@ -1,3 +1,9 @@
+package atlas.commands;
+
+import atlas.storage.Storage;
+import atlas.tasks.ItemList;
+import atlas.ui.Ui;
+
 public class ByeCommand implements Command {
     @Override
     public void execute(ItemList itemList, Ui ui, Storage storage) {
