+++ src/main/java/atlas/commands/UnmarkCommand.java
@@ -1,3 +1,9 @@
+package atlas.commands;
+
+import atlas.storage.Storage;
+import atlas.tasks.ItemList;
+import atlas.ui.Ui;
+
 public class UnmarkCommand implements Command {
     int index;
