+++ src/main/java/atlas/exceptions/EmptyTaskNameException.java
@@ -1,3 +1,5 @@
+package atlas.exceptions;
+
 public class EmptyTaskNameException extends Exception {
     public EmptyTaskNameException() {
         super("Task name cannot be empty.");
