+++ src/main/java/atlas/exceptions/InvalidDateFormatException.java
@@ -1,3 +1,5 @@
+package atlas.exceptions;
+
 public class InvalidDateFormatException extends Exception{
     public InvalidDateFormatException() {
         super("Please enter a valid date in the format dd/MM/yyyy HH:mm");
