+++ src/main/java/atlas/exceptions/InvalidDateRangeException.java
@@ -1,3 +1,5 @@
+package atlas.exceptions;
+
 public class InvalidDateRangeException extends Exception{
     public InvalidDateRangeException() {
         super("FROM date cannot be before the TO date. ");
