+++ src/main/java/atlas/exceptions/InvalidFormatDeadlineException.java
@@ -1,3 +1,5 @@
+package atlas.exceptions;
+
 public class InvalidFormatDeadlineException extends Exception {
     public InvalidFormatDeadlineException() {
         super("Format: deadline <name> /by <datetime>");
