+++ src/main/java/atlas/exceptions/InvalidFormatEventException.java
@@ -1,3 +1,5 @@
+package atlas.exceptions;
+
 public class InvalidFormatEventException extends Exception {
     public InvalidFormatEventException() {
         super("Format: event <name> /from <datetime> /to <datetime>");
