+++ src/main/java/atlas/exceptions/InvalidTaskNumberException.java
@@ -1,3 +1,5 @@
+package atlas.exceptions;
+
 public class InvalidTaskNumberException extends Exception{
     public InvalidTaskNumberException() {
         super("Please enter a valid task number.");
