+++ src/main/java/atlas/exceptions/PastDateException.java
@@ -1,3 +1,5 @@
+package atlas.exceptions;
+
 public class PastDateException extends Exception{
     public PastDateException() {
         super("Date cannot be before today.");
