+++ src/main/java/atlas/exceptions/UnknownCommandException.java
@@ -1,2 +1,4 @@
+package atlas.exceptions;
+
 public class UnknownCommandException extends Exception{
 }
