+++ src/main/java/atlas/tasks/Deadline.java
@@ -1,3 +1,6 @@
+package atlas.tasks;
+
+import atlas.utilities.Parser;
 import java.time.LocalDateTime;
 
 public class Deadline extends Item {
