+++ src/main/java/atlas/tasks/Todo.java
@@ -1,3 +1,5 @@
+package atlas.tasks;
+
 public class Todo extends Item {
 
     public Todo(String name) {
