+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ____________________________________________________________
-     Hello, I'm Atlas!
+     Hello, I'm atlas.Atlas!
      What do you want to do?
     ____________________________________________________________
     ____________________________________________________________
